Refactor StringCleanup.removeCharacters method
