+++ test/groovy/org/ods/orchestration/service/JiraServiceSpec.groovy
@@ -2353,14 +2353,4 @@ class JiraServiceSpec extends SpecHelper {
         stopServer(server)
     }
 
-    def "Remove nonbreakable white space"() {
-        given: 'String with nonbreakable white space'
-        def inputString = '\u00A0This\u00A0string\u00A0has\u00A0wrong\u00A0spaces\u00A0'
-
-        when: 'We remove the characters'
-        def outputString = JiraService.removeCharacters(inputString)
-
-        then: 'The characters were removed'
-        outputString == ' This string has wrong spaces '
-    }
 }
